Move Bootstrap Tour lib to the JS folder
