+++ webapp/src/js/bootstrap-tour-standalone.js
