+++ webapp/src/ts/main.ts
@@ -8,7 +8,7 @@ window.startupTimes.firstCodeExecution = performance.now();
 window.PouchDB = require('pouchdb-browser').default;
 window.PouchDB.plugin(require('pouchdb-debug'));
 window.$ = window.jQuery = require('jquery');
-window.Tour = require('./bootstrap-tour-standalone');
+window.Tour = require('../js/bootstrap-tour-standalone');
 //window.d3 = require('d3');
 
 //import * as Select2 from '../../node_modules/select2/dist/js/select2.full';
